+++ lib/markdown_text_input.dart
@@ -1,20 +1,17 @@
+import 'package:expandable/expandable.dart';
 import 'package:flutter/material.dart';
 import 'package:markdown_editable_textinput/format_markdown.dart';
-import 'package:expandable/expandable.dart';
 
 /// Widget with markdown buttons
 class MarkdownTextInput extends StatefulWidget {
   /// Callback called when text changed
-  final Function onTextChanged;
+  final Function(String text) onTextChanged;
 
   /// Initial value you want to display
-  final String initialValue;
+  final String? initialValue;
 
-  /// Validator for the TextFormField
+  /// `Validator` for the `TextFormField`
   final String? Function(String? value)? validators;
-
-  /// String displayed at hintText in TextFormField
-  final String? label;
 
   /// Change the text direction of the input (RTL / LTR)
   final TextDirection? textDirection;
@@ -25,14 +22,29 @@ class MarkdownTextInput extends StatefulWidget {
   /// List of action the component can handle
   final List<MarkdownType> actions;
 
-  /// Optionnal controller to manage the input
+  /// Optional `controller` to manage the input
   final TextEditingController? controller;
+
+  /// `InputDecoration` of `TextFormField`
+  final InputDecoration? inputDecoration;
+
+  /// Expands actions widget to full size
+  final bool fillActions;
+
+  /// `BoxDecoration` of actions `Container`
+  final BoxDecoration? actionsBoxDecoration;
+
+  /// Space between `actions` widget and `TextFormField` widget
+  final double? separaterHeight;
+
+  /// `Padding` of `Actions button` widget
+  final EdgeInsets actionsButtonPadding;
 
   /// Constructor for [MarkdownTextInput]
   MarkdownTextInput(
-    this.onTextChanged,
-    this.initialValue, {
-    this.label = '',
+    this.onTextChanged, {
+    this.initialValue,
+    this.controller,
     this.validators,
     this.textDirection = TextDirection.ltr,
     this.maxLines = 10,
@@ -43,40 +55,51 @@ class MarkdownTextInput extends StatefulWidget {
       MarkdownType.link,
       MarkdownType.list
     ],
-    this.controller,
+    this.inputDecoration,
+    this.fillActions = false,
+    this.actionsBoxDecoration,
+    this.separaterHeight = 15,
+    this.actionsButtonPadding = const EdgeInsets.all(15),
   });
 
   @override
-  _MarkdownTextInputState createState() => _MarkdownTextInputState(controller ?? TextEditingController());
+  _MarkdownTextInputState createState() =>
+      _MarkdownTextInputState(controller ?? TextEditingController());
 }
 
 class _MarkdownTextInputState extends State<MarkdownTextInput> {
   final TextEditingController _controller;
-  TextSelection textSelection = const TextSelection(baseOffset: 0, extentOffset: 0);
+  TextSelection textSelection =
+      const TextSelection(baseOffset: 0, extentOffset: 0);
 
   _MarkdownTextInputState(this._controller);
 
   void onTap(MarkdownType type, {int titleSize = 1}) {
     final basePosition = textSelection.baseOffset;
-    var noTextSelected = (textSelection.baseOffset - textSelection.extentOffset) == 0;
+    var noTextSelected =
+        (textSelection.baseOffset - textSelection.extentOffset) == 0;
 
-    final result = FormatMarkdown.convertToMarkdown(
-        type, _controller.text, textSelection.baseOffset, textSelection.extentOffset,
+    final result = FormatMarkdown.convertToMarkdown(type, _controller.text,
+        textSelection.baseOffset, textSelection.extentOffset,
         titleSize: titleSize);
 
-    _controller.value = _controller.value
-        .copyWith(text: result.data, selection: TextSelection.collapsed(offset: basePosition + result.cursorIndex));
+    _controller.value = _controller.value.copyWith(
+        text: result.data,
+        selection:
+            TextSelection.collapsed(offset: basePosition + result.cursorIndex));
 
     if (noTextSelected) {
-      _controller.selection = TextSelection.collapsed(offset: _controller.selection.end - result.replaceCursorIndex);
+      _controller.selection = TextSelection.collapsed(
+          offset: _controller.selection.end - result.replaceCursorIndex);
     }
   }
 
   @override
   void initState() {
-    _controller.text = widget.initialValue;
+    _controller.text = widget.initialValue ?? '';
     _controller.addListener(() {
-      if (_controller.selection.baseOffset != -1) textSelection = _controller.selection;
+      if (_controller.selection.baseOffset != -1)
+        textSelection = _controller.selection;
       widget.onTextChanged(_controller.text);
     });
     super.initState();
@@ -90,39 +113,18 @@ class _MarkdownTextInputState extends State<MarkdownTextInput> {
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      decoration: BoxDecoration(
-        color: Theme.of(context).cardColor,
-        border: Border.all(color: Theme.of(context).colorScheme.secondary, width: 2),
-        borderRadius: const BorderRadius.all(Radius.circular(10)),
-      ),
-      child: Column(
-        children: <Widget>[
-          TextFormField(
-            textInputAction: TextInputAction.newline,
-            maxLines: widget.maxLines,
-            controller: _controller,
-            textCapitalization: TextCapitalization.sentences,
-            validator: (value) => widget.validators!(value),
-            cursorColor: Theme.of(context).primaryColor,
-            textDirection: widget.textDirection,
-            decoration: InputDecoration(
-              enabledBorder:
-                  UnderlineInputBorder(borderSide: BorderSide(color: Theme.of(context).colorScheme.secondary)),
-              focusedBorder:
-                  UnderlineInputBorder(borderSide: BorderSide(color: Theme.of(context).colorScheme.secondary)),
-              hintText: widget.label,
-              hintStyle: const TextStyle(color: Color.fromRGBO(63, 61, 86, 0.5)),
-              contentPadding: const EdgeInsets.symmetric(vertical: 15, horizontal: 10),
-            ),
-          ),
-          SizedBox(
-            height: 44,
-            child: Material(
-              color: Theme.of(context).cardColor,
-              borderRadius: const BorderRadius.only(bottomLeft: Radius.circular(10), bottomRight: Radius.circular(10)),
-              child: ListView(
-                scrollDirection: Axis.horizontal,
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: <Widget>[
+        Container(
+          clipBehavior: Clip.hardEdge,
+          decoration: widget.actionsBoxDecoration,
+          width: widget.fillActions ? double.maxFinite : null,
+          child: Material(
+            color: Colors.transparent,
+            child: SingleChildScrollView(
+              scrollDirection: Axis.horizontal,
+              child: Row(
                 children: widget.actions.map((type) {
                   return type == MarkdownType.title
                       ? ExpandableNotifier(
@@ -134,36 +136,45 @@ class _MarkdownTextInputState extends State<MarkdownTextInput> {
                                   padding: EdgeInsets.all(10),
                                   child: Text(
                                     'H#',
-                                    style: TextStyle(fontSize: 16, fontWeight: FontWeight.w700),
+                                    style: TextStyle(
+                                        fontSize: 16,
+                                        fontWeight: FontWeight.w700),
                                   ),
                                 ),
                               ),
                             ),
                             expanded: Container(
                               color: Colors.white10,
-                              child: Row(
-                                children: [
-                                  for (int i = 1; i <= 6; i++)
-                                    InkWell(
-                                      key: Key('H${i}_button'),
-                                      onTap: () => onTap(MarkdownType.title, titleSize: i),
+                              child: SingleChildScrollView(
+                                scrollDirection: Axis.horizontal,
+                                physics: NeverScrollableScrollPhysics(),
+                                child: Row(
+                                  children: [
+                                    for (int i = 1; i <= 6; i++)
+                                      InkWell(
+                                        key: Key('H${i}_button'),
+                                        onTap: () => onTap(MarkdownType.title,
+                                            titleSize: i),
+                                        child: Padding(
+                                          padding: widget.actionsButtonPadding,
+                                          child: Text(
+                                            'H$i',
+                                            style: TextStyle(
+                                                fontSize: (18 - i).toDouble(),
+                                                fontWeight: FontWeight.w700),
+                                          ),
+                                        ),
+                                      ),
+                                    ExpandableButton(
                                       child: Padding(
-                                        padding: const EdgeInsets.all(10),
-                                        child: Text(
-                                          'H$i',
-                                          style: TextStyle(fontSize: (18 - i).toDouble(), fontWeight: FontWeight.w700),
+                                        padding: widget.actionsButtonPadding,
+                                        child: Icon(
+                                          Icons.close,
                                         ),
                                       ),
                                     ),
-                                  ExpandableButton(
-                                    child: const Padding(
-                                      padding: EdgeInsets.all(10),
-                                      child: Icon(
-                                        Icons.close,
-                                      ),
-                                    ),
-                                  ),
-                                ],
+                                  ],
+                                ),
                               ),
                             ),
                           ),
@@ -172,16 +183,27 @@ class _MarkdownTextInputState extends State<MarkdownTextInput> {
                           key: Key(type.key),
                           onTap: () => onTap(type),
                           child: Padding(
-                            padding: EdgeInsets.all(10),
+                            padding: widget.actionsButtonPadding,
                             child: Icon(type.icon),
                           ),
                         );
                 }).toList(),
               ),
             ),
-          )
-        ],
-      ),
+          ),
+        ),
+        SizedBox(height: widget.separaterHeight),
+        TextFormField(
+          textInputAction: TextInputAction.newline,
+          maxLines: widget.maxLines,
+          controller: _controller,
+          textCapitalization: TextCapitalization.sentences,
+          validator: (value) => widget.validators!(value),
+          cursorColor: Theme.of(context).primaryColor,
+          textDirection: widget.textDirection,
+          decoration: widget.inputDecoration,
+        ),
+      ],
     );
   }
 }